+++ lua/LUA/ak/io/AkStatistik.lua
@@ -45,7 +45,7 @@ AkStatistik.programVersion = "0.8.4"
 -- false: Update json file without checking if the EEP-Web Server is ready
 AkStatistik.checkServerStatus = true
 
-local registeredPublishers = {}
+local registeredJsonCollectors = {}
 local writeLater = {}
 local checksum = 0
 local initialized = false
@@ -90,56 +90,105 @@ local function fillApiEntriesV1(orderedKeys)
     writeLater["api-entries"] = apiEntries
 end
 
-local function initializePublisher(publisher)
+local function initializeJsonCollector(jsonCollector)
     local t0 = os.clock()
-    publisher.initialize()
+    jsonCollector.initialize()
     local t1 = os.clock()
     local timeDiff = t1 - t0
-    print(string.format('AkStatistik: initialize() %.3f Sekunden fuer "%s"', timeDiff, publisher.name))
+    print(string.format('AkStatistik: initialize() %.3f Sekunden fuer "%s"', timeDiff, jsonCollector.name))
+end
+
+local function collectFrom(jsonCollector, printFirstTime)
+    local t0 = os.clock()
+    local newData = jsonCollector.collectData()
+    local t1 = os.clock()
+    local timeDiff = t1 - t0
+    if timeDiff > 0.02 or printFirstTime then
+        print(string.format('AkStatistik:collectData() %.3f Sekunden fuer "%s"', timeDiff, jsonCollector.name))
+    end
+    return newData
+end
+
+local function collectData(printFirstTime)
+    for _, jsonCollector in pairs(registeredJsonCollectors) do
+        local newData = collectFrom(jsonCollector, printFirstTime)
+        for key, value in pairs(newData) do
+            writeLater[key] = value
+        end
+    end
 end
 
 --- Initialize data.
 -- do it once
 local function initialize()
-    local DefaultPublishers = require("ak.core.DefaultPublishers")
-    DefaultPublishers.addDefaultPublishers()
-
     print("AkStatistik: initialize()")
-    for _, publisher in pairs(registeredPublishers) do
-        initializePublisher(publisher)
+    for _, jsonCollector in pairs(registeredJsonCollectors) do
+        initializeJsonCollector(jsonCollector)
     end
 
     initialized = true
 end
 
-function AkStatistik.addPublisher(...)
-    for _, publisher in ipairs({...}) do
-        -- Check the publisher
+function AkStatistik.addJsonCollector(...)
+    for _, jsonCollector in ipairs({...}) do
+        -- Check the jsonCollector
         assert(
-            publisher.name and type(publisher.name) == "string",
+            jsonCollector.name and type(jsonCollector.name) == "string",
             "Der Name des Moduls muss gesetzt und ein String sein"
         )
         assert(
-            publisher.initialize and type(publisher.initialize) == "function",
+            jsonCollector.initialize and type(jsonCollector.initialize) == "function",
             "Das Modul muss eine Funktion initialize() besitzen"
         )
         assert(
-            publisher.updateData and type(publisher.updateData) == "function",
-            "Das Modul muss eine Funktion updateData() besitzen"
+            jsonCollector.collectData and type(jsonCollector.collectData) == "function",
+            "Das Modul muss eine Funktion collectData() besitzen"
         )
 
-        -- Remember the publisher by it's name
-        print(string.format("AkStatistik.addPublisher(%s)", publisher.name))
-        registeredPublishers[publisher.name] = publisher
+        -- Remember the jsonCollector by it's name
+        print(string.format("AkStatistik.addJsonCollector(%s)", jsonCollector.name))
+        registeredJsonCollectors[jsonCollector.name] = jsonCollector
 
         if initialized then
-            initializePublisher(publisher)
+            initializeJsonCollector(jsonCollector)
         end
     end
 end
 
+function collectAndWriteData(printFirstTime)
+    local t0 = os.clock()
+    collectData(printFirstTime)
+
+    -- add statistical data
+    local t1 = os.clock()
+    local orderedKeys = {}
+    for key in pairs(writeLater) do
+        table.insert(orderedKeys, key)
+    end
+    table.sort(orderedKeys)
+    fillApiEntriesV1(orderedKeys)
+
+    -- write file
+    local t2 = os.clock()
+    AkWebServerIo.updateJsonFile(json.encode(writeLater, {keyorder = orderedKeys}))
+
+    local t3 = os.clock()
+    if t3 - t0 > 1 then
+        print(
+            string.format(
+                "WARNUNG: AkStatistik:collectAndWriteData(): dauerte %.2f s - " ..
+                    "collect: %.2f s, sort: %.2f s, writeFile: %.2f s",
+                t3 - t0,
+                t1 - t0,
+                t2 - t1,
+                t3 - t2
+            )
+        )
+    end
+end
 
 local i = -1
+
 --- Main function of this module.
 -- Call this function in main loop of EEP.
 -- do it frequently
@@ -149,72 +198,41 @@ function AkStatistik.statistikAusgabe(modulus)
     if not modulus or type(modulus) ~= "number" then
         modulus = 5
     end
+    i = i + 1
 
-    local printFirstTime = false
     local overallTime0 = os.clock()
+    local serverIsReady = not AkStatistik.checkServerStatus or AkWebServerIo.checkWebServer()
+
     -- initialization in first call
-    if not initialized and (not AkStatistik.checkServerStatus or AkWebServerIo.checkWebServer()) then
+    local printFirstTime = false
+    if not initialized and serverIsReady then
         printFirstTime = true
         initialize()
     end
-    i = i + 1
 
     -- process commands
+    local overallTime1 = os.clock()
     AkWebServerIo.processNewCommands()
 
     -- export data regularly
-    if i % modulus == 0 and (not AkStatistik.checkServerStatus or AkWebServerIo.checkWebServer()) then
-        local t0 = os.clock()
-        for _, publisher in pairs(registeredPublishers) do
-            -- local t0 = os.clock()
-            publisher.updateData()
-            -- local t1 = os.clock()
-            -- local timeDiff = t1 - t0
-            -- print(string.format("AkStatistik: updateData() %.3f Sekunden fuer \"%s\"", timeDiff, publisher.name))
-        end
-
-        -- add statistical data
-        local orderedKeys = {}
-        for key in pairs(writeLater) do
-            table.insert(orderedKeys, key)
-        end
-        table.sort(orderedKeys)
-        fillApiEntriesV1(orderedKeys)
-
-        local t1 = os.clock()
-
-        -- export data
-        AkWebServerIo.updateJsonFile(
-            json.encode(
-                writeLater,
-                {
-                    keyorder = orderedKeys
-                }
-            )
-        )
-
-        -- clear additional data
-        -- NOT USED ANYMORE! writeLater = {}
-
-        local t2 = os.clock()
-
-        local timeDiff = t2 - t0
-        if timeDiff > 1 then
-            print(string.format("AkStatistik: fill: %.2f sec, store: %.2f sec", t1 - t0, t2 - t1))
-            print(
-                string.format(
-                    "WARNUNG: AkStatistik.statistikAusgabe schreiben dauerte %.2f Sekunden " ..
-                        "(nur interessant, wenn EEP nicht pausiert wurde)",
-                    timeDiff
-                )
-            )
-        end
+    local overallTime2 = os.clock()
+    if i % modulus == 0 and serverIsReady then
+        collectAndWriteData(printFirstTime)
     end
 
-    local overallTime1 = os.clock()
-    local timeDiff = overallTime1 - overallTime0
+    local overallTime3 = os.clock()
+    local timeDiff = overallTime3 - overallTime0
     if printFirstTime then
-        print(string.format("FIRST TIME for AkStatistik.statistikAusgabe() was %.3f s", timeDiff))
+        print(
+            string.format(
+                "FIRST TIME for AkStatistik.statistikAusgabe() was %.3f s" ..
+                    "\ninitialize: %.3f\nprocessNewCommands: %.3f\ncollectAndWriteData: %.3f",
+                timeDiff,
+                overallTime3 - overallTime2,
+                overallTime2 - overallTime1,
+                overallTime1 - overallTime0
+            )
+        )
     else
         local allowedTimeDiff = modulus * 0.200
         if timeDiff > allowedTimeDiff then